+++ apps/explorer/priv/repo/migrations/20230303125841_add_op_indexes.exs
@@ -1,8 +1,0 @@
-defmodule Explorer.Repo.Migrations.AddOpIndexes do
-  use Ecto.Migration
-
-  def change do
-    create(index(:op_output_roots, [:l1_block_number]))
-    create(index(:op_withdrawal_events, [:l1_block_number]))
-  end
-end