+++ apps/explorer/priv/repo/migrations/20231025102325_add_op_withdrawal_index.exs
@@ -1,7 +1,0 @@
-defmodule Explorer.Repo.Migrations.AddOpWithdrawalIndex do
-  use Ecto.Migration
-
-  def change do
-    create(index(:op_withdrawals, :l2_transaction_hash))
-  end
-end