+++ apps/explorer/priv/repo/migrations/20240124124644_remove_op_epoch_number_field.exs
@@ -1,9 +1,0 @@
-defmodule Explorer.Repo.Migrations.RemoveOpEpochNumberField do
-  use Ecto.Migration
-
-  def change do
-    alter table(:op_transaction_batches) do
-      remove(:epoch_number)
-    end
-  end
-end